LPS-78433 Allow the return of a false value when submitting a single unselected checkbox form with a predefined value (defaultDDMFormFieldParameterValue) set to true. The only time the value being returned is incorrect is when the form is actually being submitted with a predefined value.
This regression was introduced by LPS-73350 fb2241828d95722815bf9543967c8da2277e10be. ServletRequest.getParameter always returns null when the parameter doesn't exist. The httpServletRequest doesn't recognize the field as unselected or false when there is a predefinded (true) value. Thus when false is submitted the parameterValue is null resulting in the return of the default value or true every time.